+++ AmazonChimeSDK/AmazonChimeSDK/audiovideo/video/backgroundfilter/backgroundblur/BackgroundBlurStrength.swift
@@ -1,0 +1,16 @@
+//
+//  BackgroundBlurStrength.swift
+//  AmazonChimeSDK
+//
+//  Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+//  SPDX-License-Identifier: Apache-2.0
+//
+
+import Foundation
+
+/// Enum defining the different background blur strength options.
+@objc public enum BackgroundBlurStrength: Int {
+    case low = 10
+    case medium = 15
+    case high = 25
+}